# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Flexfence Backend API
  version: 0.1.0
  description: |
    MVP attendance API for Flexfence.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/consent-field-recommendations:
    get:
      tags: [Events]
      summary: List frequently used custom consent fields for the organization
      operationId: listConsentFieldRecommendations
      security:
        - BusinessBearer: []
      responses:
        "200":
          description: Recommendations ordered by use count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListConsentFieldRecommendationsResponse"
        default:
          $ref: "#/components/responses/APIError"
  /v1/events:
    get:
      tags: [Events]
      summary: List events
      operationId: listEvents
      security:
        - BusinessBearer: []
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEventsResponse"
        default:
          $ref: "#/components/responses/APIError"
    post:
      tags: [Events]
      summary: Create event
      operationId: createEvent
      security:
        - BusinessBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        default:
          $ref: "#/components/responses/APIError"
...
  /v1/events/{eventId}/fences:
    get:
      tags: [Events]
      summary: List fences for event
      operationId: listFences
      security:
        - BusinessBearer: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Fence list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListFencesResponse"
        default:
          $ref: "#/components/responses/APIError"
    post:
      tags: [Events]
      summary: Create fence for event
      operationId: createFence
      security:
        - BusinessBearer: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFenceRequest"
      responses:
        "201":
          description: Fence created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fence"
        default:
          $ref: "#/components/responses/APIError"
  /v1/events/{eventId}/fences/{fenceId}:
    delete:
      tags: [Events]
      summary: Delete a fence from an event
      operationId: deleteFence
      security:
        - BusinessBearer: []
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/FenceId"
      responses:
        "204":
          description: Fence deleted
        "409":
          description: Event is live; fence deletion is blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        default:
          $ref: "#/components/responses/APIError"
...
  /v1/events/{eventId}/consent-template:
    get:
      tags: [Events]
      summary: Get consent template for event
      operationId: getConsentTemplate
      security:
        - BusinessBearer: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Consent template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentTemplate"
        default:
          $ref: "#/components/responses/APIError"
    put:
      tags: [Events]
      summary: Save consent template for event
      operationId: saveConsentTemplate
      security:
        - BusinessBearer: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveConsentTemplateRequest"
      responses:
        "200":
          description: Saved consent template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentTemplate"
        default:
          $ref: "#/components/responses/APIError"
...
  /v1/events/{eventId}/analytics:
    get:
      tags: [Events]
      summary: Event-level attendance and join analytics
      operationId: getEventAnalytics
      security:
        - BusinessBearer: []
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventAnalytics"
        default:
          $ref: "#/components/responses/APIError"
  /v1/events/{eventId}/fences/{fenceId}/analytics:
    get:
      tags: [Events]
      summary: Geofence-level attendance analytics
      operationId: getFenceAnalytics
      security:
        - BusinessBearer: []
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/FenceId"
      responses:
        "200":
          description: Fence analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FenceAnalytics"
        default:
          $ref: "#/components/responses/APIError"
...
components:
  securitySchemes:
    BusinessBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dashboard business user token
    UserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Mobile end-user token
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
    FenceId:
      name: fenceId
      in: path
      required: true
      schema:
        type: string
  responses:
    APIError:
      description: Standard API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
  schemas:
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    CreateEventRequest:
      type: object
      required: [title, start_at, end_at]
      properties:
        title:
          type: string
        description:
          type: string
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
    ListEventsResponse:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
    Event:
      type: object
      required: [id, organization_id, created_by_id, title, status, created_at]
      properties:
        id:
          type: string
        organization_id:
          type: string
        created_by_id:
          type: string
        title:
          type: string
        description:
          type: string
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        status:
          type: string
        created_at:
          type: string
          format: date-time
    CreateFenceRequest:
      type: object
      required: [shape_type]
      properties:
        name:
          type: string
          description: Optional; auto-generated from event title and schedule when omitted
        shape_type:
          type: string
          enum: [circle, polygon]
        start_at:
          type: string
          format: date-time
          description: Optional; inherits event start_at when omitted
        end_at:
          type: string
          format: date-time
          description: Optional; inherits event end_at when omitted
        center_lat:
          type: number
          format: double
        center_lng:
          type: number
          format: double
        radius_m:
          type: number
          format: double
        polygon_geojson:
          type: string
    ListFencesResponse:
      type: object
      required: [fences]
      properties:
        fences:
          type: array
          items:
            $ref: "#/components/schemas/Fence"
    ConsentField:
      type: object
      required: [key, label, required]
      properties:
        key:
          type: string
        label:
          type: string
        required:
          type: boolean
        is_custom:
          type: boolean
          description: True for organization-defined fields stored on the event
        value_type:
          type: string
          enum: [text, email, phone, number, date, boolean]
          description: Input type shown to attendees (required for custom fields)
    ConsentFieldRecommendation:
      type: object
      required: [key, label, value_type, use_count]
      properties:
        key:
          type: string
        label:
          type: string
        value_type:
          type: string
        use_count:
          type: integer
    ListConsentFieldRecommendationsResponse:
      type: object
      required: [recommendations]
      properties:
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/ConsentFieldRecommendation"
    TimelinePoint:
      type: object
      required: [label, count]
      properties:
        label:
          type: string
        count:
          type: integer
    FenceAnalyticsSummary:
      type: object
      required: [fence_id, fence_name, total_attendance, unique_attendees, attendance_in_fence]
      properties:
        fence_id:
          type: string
        fence_name:
          type: string
        total_attendance:
          type: integer
        unique_attendees:
          type: integer
        attendance_in_fence:
          type: integer
    EventAnalytics:
      type: object
      required:
        - event_id
        - total_joins
        - unique_joined_users
        - total_attendance
        - unique_attendees
        - joins_by_source
        - attendance_by_source
        - attendance_timeline
        - fence_summaries
      properties:
        event_id:
          type: string
        total_joins:
          type: integer
        unique_joined_users:
          type: integer
        total_attendance:
          type: integer
        unique_attendees:
          type: integer
        joins_by_source:
          type: object
          additionalProperties:
            type: integer
        attendance_by_source:
          type: object
          additionalProperties:
            type: integer
        attendance_timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelinePoint"
        fence_summaries:
          type: array
          items:
            $ref: "#/components/schemas/FenceAnalyticsSummary"
    FenceAnalytics:
      type: object
      required:
        - fence_id
        - fence_name
        - event_id
        - total_attendance
        - unique_attendees
        - attendance_by_source
        - attendance_timeline
        - unattributed_nearby
      properties:
        fence_id:
          type: string
        fence_name:
          type: string
        event_id:
          type: string
        total_attendance:
          type: integer
        unique_attendees:
          type: integer
        attendance_by_source:
          type: object
          additionalProperties:
            type: integer
        attendance_timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelinePoint"
        unattributed_nearby:
          type: integer
...
    ConsentTemplate:
      type: object
      required: [event_id, required_fields, track_entry_exit, track_movement]
      properties:
        id:
          type: string
        event_id:
          type: string
        required_fields:
          type: array
          items:
            $ref: "#/components/schemas/ConsentField"
        track_entry_exit:
          type: boolean
        track_movement:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SaveConsentTemplateRequest:
      type: object
      required: [required_fields, track_entry_exit, track_movement]
      properties:
        required_fields:
          type: array
          items:
            $ref: "#/components/schemas/ConsentField"
        track_entry_exit:
          type: boolean
        track_movement:
          type: boolean
...
    Fence:
      type: object
      required: [id, event_id, name, shape_type, start_at, end_at, created_at]
      properties:
        id:
          type: string
        event_id:
          type: string
        name:
          type: string
        shape_type:
          type: string
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        center_lat:
          type: number
          format: double
        center_lng:
          type: number
          format: double
        radius_m:
          type: number
          format: double
        polygon_geojson:
          type: string
        created_at:
          type: string
          format: date-time
...